moved config to root
